move 'debug' commands out
