+++ src/searchd.cpp
@@ -37,7 +37,7 @@
 #include "searchdddl.h"
 #include "networking_daemon.h"
 #include "query_status.h"
-#include "sphinxql_debug.h"
+#include "debug_cmds.h"
 #include "stackmock.h"
 #include "binlog.h"
 #include "indexfiles.h"
@@ -166,7 +166,6 @@ static int				g_iShutdownTimeoutUs	= 3000000; // default timeout on daemon shutd
 static int				g_iBacklog			= SEARCHD_BACKLOG;
 static int				g_iThdQueueMax		= 0;
 static auto&			g_iTFO = sphGetTFO ();
-static CSphString		g_sShutdownToken;
 static int				g_iServerID = 0;
 static bool				g_bServerID = false;
 static bool				g_bJsonConfigLoadedOk = false;
@@ -6532,8 +6531,6 @@ static uint64_t GetIndexMass ( const CSphString & sName )
 // declared to be used in ParseSysVar
 void HandleMysqlShowThreads ( RowBuffer_i & tOut, const SqlStmt_t * pStmt );
 void HandleMysqlShowTables ( RowBuffer_i & tOut, const SqlStmt_t * pStmt );
-void HandleTasks ( RowBuffer_i & tOut );
-void HandleSched ( RowBuffer_i & tOut );
 void HandleShowSessions ( RowBuffer_i& tOut, const SqlStmt_t* pStmt );
 void HandleMysqlDescribe ( RowBuffer_i & tOut, const SqlStmt_t * pStmt );
 void HandleSelectIndexStatus ( RowBuffer_i & tOut, const SqlStmt_t * pStmt );
@@ -15027,12 +15024,6 @@ void HandleMysqlFlush ( RowBuffer_i & tOut, const SqlStmt_t & )
 	tOut.Eof();
 }
 
-// stuff for command 'debug', isolated
-inline static CSphString strSHA1 ( const CSphString& sLine )
-{
-	return CalcSHA1 ( sLine.cstr(), sLine.Length() );
-}
-
 int GetLogFD ()
 {
 	if ( g_bLogStdout && g_iLogFile!=STDOUT_FILENO )
@@ -15040,211 +15031,10 @@ int GetLogFD ()
 	return g_iLogFile;
 }
 
-bool PollOptimizeRunning ( const CSphString & sIndex )
+
+const CSphString & sphGetLogFile () noexcept
 {
-	while ( true )
-	{
-		Threads::Coro::SleepMsec ( 500 );
-		auto pTmpIndex = GetServed ( sIndex );
-		if ( !ServedDesc_t::IsMutable ( pTmpIndex ) )
-			return false;
-
-		RIdx_T<RtIndex_i *> pRtIndex { pTmpIndex };
-		if ( !pRtIndex->OptimizesRunning () )
-			return true;
-	}
-}
-
-void HandleMysqlOptimizeManual ( RowBuffer_i & tOut, const DebugCmd::DebugCommand_t & tCmd )
-{
-	if ( !sphCheckWeCanModify ( tOut ) )
-		return;
-
-	auto sIndex = tCmd.m_sParam;
-	auto pIndex = GetServed ( sIndex );
-	if ( !ServedDesc_t::IsMutable ( pIndex ) )
-	{
-		tOut.Error ( "MERGE requires an existing RT table" );
-		return;
-	}
-
-	OptimizeTask_t tTask;
-	tTask.m_eVerb = OptimizeTask_t::eMerge;
-	tTask.m_iFrom = (int)tCmd.m_iPar1;
-	tTask.m_iTo = (int)tCmd.m_iPar2;
-	tTask.m_bByOrder = !tCmd.bOpt ( "byid", session::GetOptimizeById() );
-	tTask.m_iCutoff = (int)tCmd.iOpt("cutoff");
-
-	RIdx_T<RtIndex_i *> ( pIndex )->StartOptimize ( std::move ( tTask ) );
-	if ( tCmd.bOpt ( "sync" ) && !PollOptimizeRunning ( sIndex ) )
-		tOut.Error ( "RT table went away during waiting" );
-	else
-		tOut.Ok ();
-}
-
-// command 'drop [chunk] X [from] <IDX> [option...]'
-void HandleMysqlDropManual ( RowBuffer_i & tOut, const DebugCmd::DebugCommand_t & tCmd )
-{
-	if ( !sphCheckWeCanModify ( tOut ) )
-		return;
-
-	auto sIndex = tCmd.m_sParam;
-	auto pIndex = GetServed ( sIndex );
-	if ( !ServedDesc_t::IsMutable ( pIndex ) )
-	{
-		tOut.Error ( "DROP requires an existing RT table" );
-		return;
-	}
-
-	OptimizeTask_t tTask;
-	tTask.m_eVerb = OptimizeTask_t::eDrop;
-	tTask.m_iFrom = (int)tCmd.m_iPar1;
-	tTask.m_bByOrder = !tCmd.bOpt ( "byid", session::GetOptimizeById() );
-
-	RIdx_T<RtIndex_i *> ( pIndex )->StartOptimize ( std::move ( tTask ) );
-	if ( tCmd.bOpt ( "sync" ) && !PollOptimizeRunning ( sIndex ) )
-		tOut.Error ( "RT table went away during waiting" );
-	else
-		tOut.Ok ();
-}
-
-void HandleMysqlCompress ( RowBuffer_i & tOut, const DebugCmd::DebugCommand_t & tCmd )
-{
-	if ( !sphCheckWeCanModify ( tOut ) )
-		return;
-
-	auto sIndex = tCmd.m_sParam;
-	auto pIndex = GetServed ( sIndex );
-	if ( !ServedDesc_t::IsMutable ( pIndex ) )
-	{
-		tOut.Error ( "COMPRESS requires an existing RT table" );
-		return;
-	}
-
-	OptimizeTask_t tTask;
-	tTask.m_eVerb = OptimizeTask_t::eCompress;
-	tTask.m_iFrom = (int) tCmd.m_iPar1;
-	tTask.m_bByOrder = !tCmd.bOpt ( "byid", session::GetOptimizeById() );
-
-	RIdx_T<RtIndex_i *> ( pIndex )->StartOptimize ( std::move ( tTask ) );
-	if ( tCmd.bOpt ( "sync" ) && !PollOptimizeRunning ( sIndex ) )
-		tOut.Error ( "RT table went away during waiting" );
-	else
-		tOut.Ok ();
-}
-
-void HandleMysqlDedup ( RowBuffer_i& tOut, const DebugCmd::DebugCommand_t& tCmd )
-{
-	if ( !sphCheckWeCanModify ( tOut ) )
-		return;
-
-	auto sIndex = tCmd.m_sParam;
-	auto pIndex = GetServed ( sIndex );
-	if ( !ServedDesc_t::IsMutable ( pIndex ) )
-	{
-		tOut.Error ( "DEDUP requires an existing RT table" );
-		return;
-	}
-
-	OptimizeTask_t tTask;
-	tTask.m_eVerb = OptimizeTask_t::eDedup;
-	tTask.m_iFrom = (int)tCmd.m_iPar1;
-	tTask.m_bByOrder = !tCmd.bOpt ( "byid", session::GetOptimizeById() );
-
-	RIdx_T<RtIndex_i*> ( pIndex )->Optimize ( std::move ( tTask ) );
-	tOut.Ok();
-}
-
-// command 'split <IDX> [chunk] N on @uservar [option...]'
-// IDX is tCmd.m_sParam
-// chunk is tCmd.m_iPar1
-// uservar is tCmd.m_sParam2
-void HandleMysqlSplit ( RowBuffer_i & tOut, const DebugCmd::DebugCommand_t & tCmd )
-{
-	if ( !sphCheckWeCanModify ( tOut ) )
-		return;
-
-	// check index existance
-	auto sIndex = tCmd.m_sParam;
-	auto pIndex = GetServed ( sIndex );
-	if ( !ServedDesc_t::IsMutable ( pIndex ) )
-	{
-		tOut.Error ( "SPLIT requires an existing RT table" );
-		return;
-	}
-
-	bool bVarFound = false;
-	IterateUservars ( [&tCmd, &bVarFound] ( const NamedRefVectorPair_t & dVar ) {
-		if ( dVar.first == tCmd.m_sParam2
-//			&& dVar.second.m_eType==USERVAR_INT_SET_TMP // uncomment this to split only by session (result of delete .. store) variables
-		)
-			bVarFound = true;
-	} );
-
-	if ( !bVarFound )
-	{
-		tOut.Error ( "SPLIT requires an existing session @uservar" );
-		return;
-	}
-
-	OptimizeTask_t tTask;
-	tTask.m_eVerb = OptimizeTask_t::eSplit;
-	tTask.m_iFrom = (int)tCmd.m_iPar1;
-	tTask.m_sUvarFilter = tCmd.m_sParam2;
-	tTask.m_bByOrder = !tCmd.bOpt ( "byid", session::GetOptimizeById() );
-
-	RIdx_T<RtIndex_i *> ( pIndex )->StartOptimize ( std::move ( tTask ) );
-	if ( tCmd.bOpt ( "sync" ) && !PollOptimizeRunning ( sIndex ) )
-		tOut.Error ( "RT table went away during waiting" );
-	else
-		tOut.Ok ();
-}
-
-
-void HandleMysqlDebugMeta ( RowBuffer_i & tOut, const DebugCmd::DebugCommand_t & tCmd, const QueryProfile_c & tProfile )
-{
-	VectorLike tLike ( tCmd.sOpt ( "like" ) );
-	tLike.MatchTupletf ( "pseudo_shards", "%d", tProfile.m_iPseudoShards );
-	tLike.MatchTupletf ( "max_matches", "%d", tProfile.m_iMaxMatches );
-	tOut.DataTable(tLike);
-}
-
-
-void HandleMysqlfiles ( RowBuffer_i & tOut, const DebugCmd::DebugCommand_t & tCmd )
-{
-	auto sIndex = tCmd.m_sParam;
-	auto pIndex = GetServed ( sIndex );
-	if ( !ServedDesc_t::IsLocal ( pIndex ) )
-	{
-		tOut.Error ( "FILES requires an existing local table" );
-		return;
-	}
-
-	StrVec_t dFiles;
-	StrVec_t dExt;
-	RIdx_c ( pIndex )->GetIndexFiles ( dFiles, dExt );
-
-	VectorLike dOut ( 0 );
-	dOut.SetColNames ( { "file" } );
-
-	auto sFormat = tCmd.sOpt ( "format" );
-	if ( sFormat!="external" )
-		dFiles.Apply ( [&dOut] ( const CSphString & a ) { dOut.Add ( a ); } );
-
-	if ( sFormat=="all" || sFormat=="external" )
-	{
-		dExt.Uniq ();
-		dExt.Apply ( [&dOut] ( const CSphString & a ) { dOut.Add ( a ); } );
-	}
-
-	tOut.DataTable ( dOut );
-}
-
-void HandleMysqlclose ( RowBuffer_i & tOut )
-{
-	auto iSocket = session::Info().GetSocket();
-	if ( iSocket >= 0 )
-		sphSockClose ( iSocket );
+	return g_sLogFile;
 }
 
 // same for select ... from index.files
@@ -15293,326 +15083,6 @@ void HandleSelectFiles ( RowBuffer_i & tOut, const SqlStmt_t * pStmt )
 		}
 	}
 	tOut.Eof();
-}
-
-void HandleShutdownCrash ( RowBuffer_i & tOut, const CSphString & sPasswd, DebugCmd::Cmd_e eCmd )
-{
-	const char * szCmd = DebugCmd::dCommands[(BYTE) eCmd].m_szExample;
-	if ( g_sShutdownToken.IsEmpty () )
-	{
-		tOut.Error ( "shutdown_token is empty. Provide it in searchd config section." );
-		return;
-	}
-
-	if ( strSHA1 ( sPasswd )!=g_sShutdownToken )
-	{
-		tOut.Error ( "FAIL" );
-		return;
-	}
-
-	tOut.HeadTuplet ( "command", "result" );
-	tOut.DataTuplet ( szCmd, "SUCCESS" );
-	tOut.Eof ();
-	if ( eCmd==DebugCmd::Cmd_e::SHUTDOWN )
-	{
-		sigterm ( 1 );
-	} else // crash
-	{
-		BYTE * pSegv = (BYTE *) ( 0 );
-		*pSegv = 'a';
-	}
-}
-
-#if !_WIN32
-void HandleProcDump ( RowBuffer_i & tOut )
-{
-	tOut.HeadTuplet ( "command", "result" );
-	if ( g_iParentPID<=0 )
-		tOut.DataTuplet ( "procdump", "Unavailable (no watchdog)" );
-	else
-	{
-		kill ( g_iParentPID, SIGUSR1 );
-		tOut.DataTupletf ( "procdump", "Sent USR1 to wathcdog (%d)", g_iParentPID );
-	}
-	tOut.Eof ();
-}
-
-void HandleGdbStatus ( RowBuffer_i & tOut )
-{
-	tOut.HeadTuplet ( "command", "result" );
-	const auto & g_bSafeGDB = getSafeGDB ();
-	if ( g_iParentPID>0 )
-		tOut.DataTupletf ( "setgdb", "Enabled, managed by watchdog (pid=%d)", g_iParentPID );
-	else if ( g_bSafeGDB )
-		tOut.DataTupletf ( "setgdb", "Enabled, managed locally because of jemalloc", g_iParentPID );
-	else if ( g_iParentPID==-1 )
-		tOut.DataTuplet ( "setgdb", "Enabled locally, MAY HANG!" );
-	else
-		tOut.DataTuplet ( "setgdb", "Disabled" );
-	tOut.Eof ();
-}
-
-void HandleSetGdb ( RowBuffer_i & tOut, bool bParam )
-{
-	tOut.HeadTuplet ( "command", "result" );
-	const auto & g_bSafeGDB = getSafeGDB ();
-
-	if ( g_iParentPID>0 )
-		tOut.DataTupletf ( "setgdb", "Enabled by watchdog (pid=%d)", g_iParentPID );
-	else if ( g_bSafeGDB )
-		tOut.DataTuplet ( "setgdb", "Enabled locally because of jemalloc" );
-	else if ( bParam )
-	{
-		g_iParentPID = -1;
-		tOut.DataTuplet ( "setgdb", "Ok, enabled locally, MAY HANG!" );
-	} else if ( !bParam )
-	{
-		g_iParentPID = 0;
-		tOut.DataTuplet ( "setgdb", "Ok, disabled" );
-	}
-	tOut.Eof ();
-}
-
-void HandleWait ( RowBuffer_i& tOutBuf, const DebugCmd::DebugCommand_t& tCmd )
-{
-	auto iTimeoutS = tCmd.iOpt ( "timeout" );
-	auto sCluster = tCmd.m_sParam;
-	auto iTime = -sphMicroTimer();
-	auto sState = WaitClusterReady ( sCluster, iTimeoutS );
-	iTime += sphMicroTimer();
-	VectorLike tOut { tCmd.sOpt ( "like" ) };
-	tOut.SetColName("name");
-	tOut.MatchTuplet ( "cluster", sCluster.cstr() );
-	tOut.MatchTuplet ( "state", sState.cstr() );
-	tOut.MatchTupletf ( "time", "%.2t", iTime );
-	tOutBuf.DataTable ( tOut );
-}
-
-void HandleWaitStatus ( RowBuffer_i& tOutBuf, const DebugCmd::DebugCommand_t& tCmd )
-{
-	auto iTimeoutS = tCmd.iOpt ( "timeout" );
-	auto sCluster = tCmd.m_sParam;
-	auto iTxn = (int)tCmd.m_iPar1;
-	auto iTime = -sphMicroTimer();
-	auto tAchieved = WaitClusterCommit ( sCluster, iTxn, iTimeoutS );
-	iTime += sphMicroTimer();
-	VectorLike tOut { tCmd.sOpt ( "like" ) };
-	tOut.SetColName ( "name" );
-	tOut.MatchTuplet ( "cluster", sCluster.cstr() );
-	tOut.MatchTupletf ( "wanted", "%d", iTxn );
-	if ( tAchieved.first>=0 )
-		tOut.MatchTupletf ( "state", "%d", tAchieved.first );
-	else
-		tOut.MatchTuplet ( "achieved", tAchieved.second.cstr() );
-	tOut.MatchTupletf ( "time", "%.2t", iTime );
-	tOutBuf.DataTable ( tOut );
-}
-#endif
-
-void HandleTrace ( RowBuffer_i& tOut, const DebugCmd::DebugCommand_t& tCmd )
-{
-	tOut.HeadTuplet ( "command", "result" );
-#ifdef PERFETTO
-	if ( tCmd.m_sParam.IsEmpty() )
-	{
-		if ( !tCmd.m_iPar1 )
-		{
-			Tracer::Stop();
-		}
-	} else
-	{
-		Tracer::Start ( tCmd.m_sParam, tCmd.m_iPar1 );
-	}
-	tOut.DataTuplet ( "debug trace ...", "SUCCESS" );
-#else
-	tOut.DataTuplet ( "debug trace ...", "FAIL, need to rebuild with Perfetto, look to src/perfetto/README.txt" );
-#endif
-	tOut.Eof();
-}
-
-void HandleToken ( RowBuffer_i & tOut, const CSphString & sParam )
-{
-	auto sSha = strSHA1 ( sParam );
-	tOut.HeadTuplet ( "command", "result" );
-	tOut.DataTuplet ( "debug token", sSha.cstr () );
-	tOut.Eof ();
-}
-
-void HandleCurl ( RowBuffer_i & tOut, const CSphString & sParam )
-{
-	auto sRes = FetchUrl ( sParam );
-	tOut.HeadTuplet ( "command", "result" );
-	tOut.DataTuplet ( "curl", sRes.cstr() );
-	tOut.Eof();
-}
-
-
-void HandlePause ( RowBuffer_i & tOut, const DebugCmd::DebugCommand_t & tCmd )
-{
-	tOut.HeadTuplet ( "command", "result" );
-	auto bPause = tCmd.m_iPar1!=0;
-	PauseAt ( tCmd.m_sParam, bPause );
-	tOut.DataTuplet ( "debug pause ...", bPause ? "Set" : "Unset" );
-	tOut.Eof ();
-}
-
-#if HAVE_MALLOC_STATS
-void HandleMallocStats ( RowBuffer_i & tOut, const CSphString& sParam )
-{
-	tOut.HeadTuplet ( "command", "result" );
-	// check where is stderr...
-	int iOldErr = ::dup ( STDERR_FILENO );
-	::dup2 ( GetLogFD (), STDERR_FILENO );
-	sphMallocStats ( sParam.cstr() );
-	::close ( STDERR_FILENO );
-	::dup2 ( iOldErr, STDERR_FILENO );
-	::close ( iOldErr );
-	tOut.DataTuplet ( "malloc_stats", g_sLogFile.cstr () );
-	tOut.Eof ();
-}
-#endif
-
-#if HAVE_MALLOC_TRIM
-void HandleMallocTrim ( RowBuffer_i & tOut )
-{
-	tOut.HeadTuplet ( "command", "result" );
-	CSphString sResult;
-	sResult.SetSprintf ( "%d", PerformMallocTrim ( 0 ) );
-	tOut.DataTuplet ( "malloc_trim", sResult.cstr () );
-	tOut.Eof ();
-}
-#endif
-
-void HandleSleep ( RowBuffer_i & tOut, int64_t iParam )
-{
-	int64_t tmStart = sphMicroTimer ();
-	Threads::Coro::SleepMsec ( Max ( iParam/1000, 1 ) );
-	int64_t tmDelta = sphMicroTimer ()-tmStart;
-
-	tOut.HeadTuplet ( "command", "result" );
-	CSphString sResult;
-	sResult.SetSprintf ( "%.3f", (float) tmDelta / 1000000.0f );
-	tOut.DataTuplet ( "sleep", sResult.cstr () );
-	tOut.Eof ();
-}
-
-void HandleTasks ( RowBuffer_i & tOut )
-{
-	if (!tOut.HeadOfStrings ( { "Name", "MaxRunners", "CurrentRunners", "TotalSpent", "LastFinished", "Executed", "Dropped", "Enqueued" } ))
-		return;
-
-	auto dTasks = TaskManager::GetTaskInfo ();
-	for ( const auto & dTask : dTasks )
-	{
-		tOut.PutString ( dTask.m_sName );
-		if ( dTask.m_iMaxRunners > 0 )
-			tOut.PutNumAsString ( dTask.m_iMaxRunners );
-		else
-			tOut.PutString ( "unlimited" );
-		tOut.PutNumAsString ( dTask.m_iCurrentRunners );
-		tOut.PutTimeAsString ( dTask.m_iTotalSpent );
-		tOut.PutTimestampAsString ( dTask.m_iLastFinished );
-		tOut.PutNumAsString ( dTask.m_iTotalRun );
-		tOut.PutNumAsString ( dTask.m_iTotalDropped );
-		tOut.PutNumAsString ( dTask.m_iAllRunners );
-		if ( !tOut.Commit () )
-			return;
-	}
-	tOut.Eof ();
-}
-
-void HandleSched ( RowBuffer_i & tOut )
-{
-	if (!tOut.HeadOfStrings ( { "Time rest", "Task" } ))
-		return;
-	auto dTasks = sph::GetSchedInfo ();
-	for ( auto& dTask : dTasks )
-	{
-		tOut.PutTimestampAsString ( dTask.m_iTimeoutStamp );
-		tOut.PutString ( dTask.m_sTask );
-		if (!tOut.Commit ())
-			return;
-	}
-	tOut.Eof ();
-}
-
-void HandleMysqlDebug ( RowBuffer_i &tOut, const DebugCmd::DebugCommand_t* pCommand, const QueryProfile_c & tProfile )
-{
-	using namespace DebugCmd;
-	bool bVipConn = session::GetVip ();
-	assert ( pCommand->Valid() );
-
-	const auto& tCmd = *pCommand;
-
-	if ( bVipConn )
-	{
-		switch ( tCmd.m_eCommand )
-		{
-		case Cmd_e::SHUTDOWN:
-		case Cmd_e::CRASH: HandleShutdownCrash ( tOut, tCmd.m_sParam, tCmd.m_eCommand ); return;
-#if !_WIN32
-		case Cmd_e::PROCDUMP: HandleProcDump ( tOut ); return;
-		case Cmd_e::SETGDB: HandleSetGdb ( tOut, tCmd.m_iPar1!=0 ); return;
-		case Cmd_e::GDBSTATUS: HandleGdbStatus ( tOut ); return;
-#endif
-		default: break;
-		}
-	}
-
-
-	switch ( tCmd.m_eCommand )
-	{
-#if HAVE_MALLOC_STATS
-	case Cmd_e::MALLOC_STATS: HandleMallocStats ( tOut, tCmd.m_sParam ); return;
-#endif
-
-#if HAVE_MALLOC_TRIM
-	case Cmd_e::MALLOC_TRIM: HandleMallocTrim ( tOut ); return;
-#endif
-	case Cmd_e::TOKEN: HandleToken ( tOut, tCmd.m_sParam ); return;
-	case Cmd_e::SLEEP: HandleSleep ( tOut, tCmd.m_iPar1 ); return;
-	case Cmd_e::TASKS: HandleTasks ( tOut ); return;
-	case Cmd_e::SCHED: HandleSched ( tOut ); return;
-	case Cmd_e::MERGE: HandleMysqlOptimizeManual ( tOut, tCmd ); return;
-	case Cmd_e::DROP: HandleMysqlDropManual ( tOut, tCmd ); return;
-	case Cmd_e::FILES: HandleMysqlfiles ( tOut, tCmd ); return;
-	case Cmd_e::CLOSE: HandleMysqlclose ( tOut ); return;
-	case Cmd_e::COMPRESS: HandleMysqlCompress ( tOut, tCmd ); return;
-	case Cmd_e::DEDUP: HandleMysqlDedup ( tOut, tCmd ); return;
-	case Cmd_e::SPLIT: HandleMysqlSplit ( tOut, tCmd ); return;
-	case Cmd_e::META: HandleMysqlDebugMeta ( tOut, tCmd, tProfile ); return;
-#if !_WIN32
-	case Cmd_e::WAIT: HandleWait ( tOut, tCmd ); return;
-	case Cmd_e::WAIT_STATUS: HandleWaitStatus ( tOut, tCmd ); return;
-#endif
-	case Cmd_e::TRACE: HandleTrace ( tOut, tCmd );	return;
-	case Cmd_e::CURL: HandleCurl ( tOut, tCmd.m_sParam ); return;
-	case Cmd_e::PAUSE: HandlePause ( tOut, tCmd ); return;
-	default: break;
-	}
-
-	// no known command; provide short help.
-	BYTE uMask = bVipConn ? DebugCmd::NEED_VIP : DebugCmd::NONE;
-#if !_WIN32
-	uMask |= DebugCmd::NO_WIN;
-#endif
-
-#if HAVE_MALLOC_STATS
-	uMask |= DebugCmd::MALLOC_STATS;
-#endif
-
-#if HAVE_MALLOC_TRIM
-	uMask |= DebugCmd::MALLOC_TRIM;
-#endif
-
-	// display a short help
-	tOut.HeadTuplet ( "command", "meaning" );
-	tOut.DataTuplet ( "flush logs", "emulate USR1 signal" );
-	tOut.DataTuplet ( "reload tables", "emulate HUP signal" );
-	for ( const auto& dCommand : DebugCmd::dCommands )
-		if ( ( dCommand.m_uTraits & uMask )==dCommand.m_uTraits )
-			tOut.DataTuplet ( dCommand.m_szExample, dCommand.m_szExplanation );
-	tOut.Eof ();
 }
 
 // fwd
@@ -20434,7 +19904,7 @@ void ConfigureSearchd ( const CSphConfig & hConf, bool bOptPIDFile, bool bTestMo
 		SetJoinCacheSize ( hSearchd.GetSize64 ( "join_cache_size", GetJoinCacheSize() ) );
 
 	// sha1 password hash for shutdown action
-	g_sShutdownToken = hSearchd.GetStr ("shutdown_token");
+	SetShutdownToken ( hSearchd.GetStr ( "shutdown_token" ) );
 
 	if ( !g_bSeamlessRotate && MutableIndexSettings_c::GetDefaults().m_bPreopen && !bTestMode )
 		sphWarning ( "preopen_indexes=1 has no effect with seamless_rotate=0" );
